Divide the RPC-related interfaces for P2P and Centralization architectures
