+++ mooncake-store/include/master_client.h
@@ -1,5 +1,10 @@
 #pragma once
 
+#include <async_simple/coro/FutureAwaiter.h>
+#include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
+#include <glog/logging.h>
+
 #include <memory>
 #include <string>
 #include <vector>
@@ -16,6 +21,9 @@
 
 namespace mooncake {
 
+template <auto Method>
+struct RpcNameTraits;
+
 static const std::string kDefaultMasterAddress = "localhost:50051";
 
 /**
@@ -23,20 +31,7 @@ static const std::string kDefaultMasterAddress = "localhost:50051";
  */
 class MasterClient {
    public:
-    MasterClient(const UUID& client_id, MasterClientMetric* metrics = nullptr)
-        : client_id_(client_id), metrics_(metrics) {
-        coro_io::client_pool<coro_rpc::coro_rpc_client>::pool_config
-            pool_conf{};
-        const char* value = std::getenv("MC_RPC_PROTOCOL");
-        if (value && std::string_view(value) == "rdma") {
-            pool_conf.client_config.socket_config =
-                coro_io::ib_socket_t::config_t{};
-        }
-        client_pools_ =
-            std::make_shared<coro_io::client_pools<coro_rpc::coro_rpc_client>>(
-                pool_conf);
-    }
-    ~MasterClient();
+    virtual ~MasterClient() = default;
 
     MasterClient(const MasterClient&) = delete;
     MasterClient& operator=(const MasterClient&) = delete;
@@ -130,67 +125,6 @@ class MasterClient {
     BatchGetReplicaList(const std::vector<std::string>& object_keys);
 
     /**
-     * @brief Starts a put operation
-     * @param key Object key
-     * @param slice_lengths Vector of slice lengths
-     * @param value_length Total value length
-     * @param config Replication configuration
-     * @return tl::expected<std::vector<Replica::Descriptor>, ErrorCode>
-     * indicating success/failure
-     */
-    [[nodiscard]] tl::expected<std::vector<Replica::Descriptor>, ErrorCode>
-    PutStart(const std::string& key, const std::vector<size_t>& slice_lengths,
-             const ReplicateConfig& config);
-
-    /**
-     * @brief Starts a batch of put operations for N objects
-     * @param keys Vector of object key
-     * @param value_lengths Vector of total value lengths
-     * @param slice_lengths Vector of vectors of slice lengths
-     * @param config Replication configuration
-     * @return ErrorCode indicating success/failure
-     */
-    [[nodiscard]] std::vector<
-        tl::expected<std::vector<Replica::Descriptor>, ErrorCode>>
-    BatchPutStart(const std::vector<std::string>& keys,
-                  const std::vector<std::vector<uint64_t>>& slice_lengths,
-                  const ReplicateConfig& config);
-
-    /**
-     * @brief Ends a put operation
-     * @param key Object key
-     * @param replica_type Type of replica (memory or disk)
-     * @return tl::expected<void, ErrorCode> indicating success/failure
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> PutEnd(
-        const std::string& key, ReplicaType replica_type);
-
-    /**
-     * @brief Ends a put operation for a batch of objects
-     * @param keys Vector of object keys
-     * @return ErrorCode indicating success/failure
-     */
-    [[nodiscard]] std::vector<tl::expected<void, ErrorCode>> BatchPutEnd(
-        const std::vector<std::string>& keys);
-
-    /**
-     * @brief Revokes a put operation
-     * @param key Object key
-     * @param replica_type Type of replica (memory or disk)
-     * @return tl::expected<void, ErrorCode> indicating success/failure
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> PutRevoke(
-        const std::string& key, ReplicaType replica_type);
-
-    /**
-     * @brief Revokes a put operation for a batch of objects
-     * @param keys Vector of object keys
-     * @return ErrorCode indicating success/failure
-     */
-    [[nodiscard]] std::vector<tl::expected<void, ErrorCode>> BatchPutRevoke(
-        const std::vector<std::string>& keys);
-
-    /**
      * @brief Removes an object and all its replicas
      * @param key Key to remove
      * @return tl::expected<void, ErrorCode> indicating success/failure
@@ -213,25 +147,6 @@ class MasterClient {
     [[nodiscard]] tl::expected<long, ErrorCode> RemoveAll();
 
     /**
-     * @brief Registers a segment to master for allocation
-     * @param segment Segment to register
-     * @return tl::expected<void, ErrorCode> indicating success/failure
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> MountSegment(
-        const Segment& segment);
-
-    /**
-     * @brief Re-mount segments, invoked when the client is the first time to
-     * connect to the master or the client Ping TTL is expired and need
-     * to remount. This function is idempotent. Client should retry if the
-     * return code is not ErrorCode::OK.
-     * @param segments Segments to remount
-     * @return tl::expected<void, ErrorCode> indicating success/failure
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> ReMountSegment(
-        const std::vector<Segment>& segments);
-
-    /**
      * @brief Unregisters a memory segment from master
      * @param segment_id ID of the segment to unmount
      * @return tl::expected<void, ErrorCode> indicating success/failure
@@ -240,51 +155,26 @@ class MasterClient {
         const UUID& segment_id);
 
     /**
-     * @brief Gets the cluster ID for the current client to use as subdirectory
-     * name
-     * @return GetClusterIdResponse containing the cluster ID
-     */
-    [[nodiscard]] tl::expected<std::string, ErrorCode> GetFsdir();
-
-    [[nodiscard]] tl::expected<GetStorageConfigResponse, ErrorCode>
-    GetStorageConfig();
-
-    /**
      * @brief Pings master to check its availability
      * @return tl::expected<PingResponse, ErrorCode>
      * containing view version and client status
      */
     [[nodiscard]] tl::expected<PingResponse, ErrorCode> Ping();
 
-    /**
-     * @brief Mounts a local disk segment into the master.
-     * @param enable_offloading If true, enables offloading (write-to-file).
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> MountLocalDiskSegment(
-        const UUID& client_id, bool enable_offloading);
-
-    /**
-     * @brief Heartbeat call to collect object-level statistics and retrieve the
-     * set of non-persisted objects.
-     * @param enable_offloading Indicates whether persistence is enabled for
-     * this segment.
-     */
-    [[nodiscard]] tl::expected<std::unordered_map<std::string, int64_t>,
-                               ErrorCode>
-    OffloadObjectHeartbeat(const UUID& client_id, bool enable_offloading);
-
-    /**
-     * @brief Adds multiple new objects to a specified client in batch.
-     * @param keys         A list of object keys (names) that were successfully
-     * offloaded.
-     * @param metadatas    The corresponding metadata for each offloaded object,
-     * including size, storage location, etc.
-     */
-    [[nodiscard]] tl::expected<void, ErrorCode> NotifyOffloadSuccess(
-        const UUID& client_id, const std::vector<std::string>& keys,
-        const std::vector<StorageObjectMetadata>& metadatas);
-
-   private:
+   protected:
+    MasterClient(const UUID& client_id, MasterClientMetric* metrics = nullptr)
+        : client_id_(client_id), metrics_(metrics) {
+        coro_io::client_pool<coro_rpc::coro_rpc_client>::pool_config
+            pool_conf{};
+        const char* value = std::getenv("MC_RPC_PROTOCOL");
+        if (value && std::string_view(value) == "rdma") {
+            pool_conf.client_config.socket_config =
+                coro_io::ib_socket_t::config_t{};
+        }
+        client_pools_ =
+            std::make_shared<coro_io::client_pools<coro_rpc::coro_rpc_client>>(
+                pool_conf);
+    }
     /**
      * @brief Generic RPC invocation helper for single-result operations
      * @tparam ServiceMethod Pointer to WrappedMasterService member function
@@ -295,7 +185,44 @@ class MasterClient {
      */
     template <auto ServiceMethod, typename ReturnType, typename... Args>
     [[nodiscard]] tl::expected<ReturnType, ErrorCode> invoke_rpc(
-        Args&&... args);
+        Args&&... args) {
+        auto pool = client_accessor_.GetClientPool();
+
+        // Increment RPC counter
+        if (metrics_) {
+            metrics_->rpc_count.inc({RpcNameTraits<ServiceMethod>::value});
+        }
+
+        auto start_time = std::chrono::steady_clock::now();
+        return async_simple::coro::syncAwait(
+            [&]() -> async_simple::coro::Lazy<
+                      tl::expected<ReturnType, ErrorCode>> {
+                auto ret = co_await pool->send_request(
+                    [&](coro_io::client_reuse_hint,
+                        coro_rpc::coro_rpc_client& client) {
+                        return client.send_request<ServiceMethod>(
+                            std::forward<Args>(args)...);
+                    });
+                if (!ret.has_value()) {
+                    LOG(ERROR) << "Client not available";
+                    co_return tl::make_unexpected(ErrorCode::RPC_FAIL);
+                }
+                auto result = co_await std::move(ret.value());
+                if (!result) {
+                    LOG(ERROR) << "RPC call failed: " << result.error().msg;
+                    co_return tl::make_unexpected(ErrorCode::RPC_FAIL);
+                }
+                if (metrics_) {
+                    auto end_time = std::chrono::steady_clock::now();
+                    auto latency =
+                        std::chrono::duration_cast<std::chrono::microseconds>(
+                            end_time - start_time);
+                    metrics_->rpc_latency.observe(
+                        {RpcNameTraits<ServiceMethod>::value}, latency.count());
+                }
+                co_return result->result();
+            }());
+    }
 
     /**
      * @brief Generic RPC invocation helper for batch operations
@@ -308,13 +235,60 @@ class MasterClient {
      */
     template <auto ServiceMethod, typename ResultType, typename... Args>
     [[nodiscard]] std::vector<tl::expected<ResultType, ErrorCode>>
-    invoke_batch_rpc(size_t input_size, Args&&... args);
+    invoke_batch_rpc(size_t input_size, Args&&... args) {
+        auto pool = client_accessor_.GetClientPool();
+
+        // Increment RPC counter
+        if (metrics_) {
+            metrics_->rpc_count.inc({RpcNameTraits<ServiceMethod>::value});
+        }
+
+        auto start_time = std::chrono::steady_clock::now();
+        return async_simple::coro::syncAwait(
+            [&]() -> async_simple::coro::Lazy<
+                      std::vector<tl::expected<ResultType, ErrorCode>>> {
+                auto ret = co_await pool->send_request(
+                    [&](coro_io::client_reuse_hint,
+                        coro_rpc::coro_rpc_client& client) {
+                        return client.send_request<ServiceMethod>(
+                            std::forward<Args>(args)...);
+                    });
+                if (!ret.has_value()) {
+                    LOG(ERROR) << "Client not available";
+                    co_return std::vector<tl::expected<ResultType, ErrorCode>>(
+                        input_size, tl::make_unexpected(ErrorCode::RPC_FAIL));
+                }
+                auto result = co_await std::move(ret.value());
+                if (!result) {
+                    LOG(ERROR)
+                        << "Batch RPC call failed: " << result.error().msg;
+                    std::vector<tl::expected<ResultType, ErrorCode>>
+                        error_results;
+                    error_results.reserve(input_size);
+                    for (size_t i = 0; i < input_size; ++i) {
+                        error_results.emplace_back(
+                            tl::make_unexpected(ErrorCode::RPC_FAIL));
+                    }
+                    co_return error_results;
+                }
+                if (metrics_) {
+                    auto end_time = std::chrono::steady_clock::now();
+                    auto latency =
+                        std::chrono::duration_cast<std::chrono::microseconds>(
+                            end_time - start_time);
+                    metrics_->rpc_latency.observe(
+                        {RpcNameTraits<ServiceMethod>::value}, latency.count());
+                }
+                co_return result->result();
+            }());
+    }
 
     /**
      * @brief Accessor for the coro_rpc_client pool. Since coro_rpc_client pool
      * cannot reconnect to a different address, a new coro_rpc_client pool is
      * created if the address is different from the current one.
      */
+   protected:
     class RpcClientAccessor {
        public:
         void SetClientPool(
@@ -335,6 +309,8 @@ class MasterClient {
         std::shared_ptr<coro_io::client_pool<coro_rpc::coro_rpc_client>>
             client_pool_;
     };
+
+   protected:
     RpcClientAccessor client_accessor_;
 
     // The client identification.
